+++ src/lib.rs
@@ -6,3 +6,4 @@
 
 pub mod gadget;
 pub mod license;
+pub mod state;